+++ app.vue
@@ -1,0 +1,10 @@
+<template>
+  <NuxtPage />
+  <UNotifications />
+</template>
+
+<style>
+body {
+  font-family: 'Baloo 2', sans-serif;
+}
+</style>