+++ nuxt.config.ts
@@ -2,5 +2,27 @@
 export default defineNuxtConfig({
   compatibilityDate: '2024-11-01',
   devtools: { enabled: true },
-  modules: ['@nuxtjs/tailwindcss', '@nuxt/fonts', '@nuxt/icon']
+  modules: ['@nuxt/ui', '@nuxtjs/tailwindcss', '@nuxt/fonts', '@nuxt/icon'],
+  colorMode: {
+    preference: 'light',
+  },
+  runtimeConfig: {
+    public: {
+      API_URL: process.env.NUXT_PUBLIC_API_URL
+    }
+  },
+  app: {
+    head: {
+      title: 'ObjLab',
+      meta: [
+        { charset: 'utf-8' },
+        { name: 'viewport', content: 'width=device-width, initial-scale=1' },
+        { hid: 'description', name: 'description', content: '' },
+        { name: 'format-detection', content: 'telephone=no' },
+      ],
+      link: [
+        { rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' },
+      ],
+    }
+  }
 })