+++ pages/signin.vue
@@ -1,0 +1,9 @@
+<template>
+  <div class="flex flex-col mt-32">  
+    <SignInForm />
+  </div>
+</template>
+
+<script setup lang="ts">
+import SignInForm from '../components/auth/SignInForm.vue';
+</script>