+++ pages/signup.vue
@@ -1,0 +1,9 @@
+<template>
+  <div class="flex flex-col mt-32">  
+    <SignUpForm />
+  </div>
+</template>
+
+<script setup lang="ts">
+import SignUpForm from '../components/auth/SignUpForm.vue';
+</script>